AD-HOC feat (deploy): Add a basic helm chart
Currently, there is no easy way to deploy the built jiralerts container
into a Kubernetes cluster. This commit adds such a method, through the
use of a "helm" chart.

A helm chart is a specification of Kubernetes resources, but allows them
to be configured slightly for a user specific implementation.

This commit adds such a chart.