AD-HOC feat (Makefile): Add a makefile
There are certain tasks that are laborious, but will be repeated often.
To assist with these tasks, a file called a "Makefile" is created. This
tool can be used on Linux system with the tool "gnu make" to automate
these tasks.

This commit adds such a makefile:

Design Notes:

pretty colours:

  The makefile is self-documenting, and explains the purpose of each
  target when "make help" (or just make) is run.