AD-HOC feat (Docs): Add a Changelog
Generally speaking, Software will be used for an extended period of
time. Further, it's also true that software is not stable, and will be
subject to change over time. There are a couple of systems to deal with
this change:

1. Git logs; useful for per change justification (such as this), as well
   as providing context useful for debugging or changing designs in
   future, and
2. The Changelog; useful for giving software consumers a summary of the
   important project changes they should be aware of upgrading.

This commits adds the changelog, which should be kept up to date with
all changes made in the repository.