Issue-5 refactor (script): Disable reopen behaviour by default
Currently, there exists a behaviour in which tickets that are closed
might be reopened. This becomes problematic, as generally speaking once
tickets are moved to a finalised status, they are never revisited again.

It is the authors view that each alert dispatch must be handled
separately. If an alert is inherently "flaky", the alert must be
modified such that it becomes less so (And thus less spurious issues are
filed)

Design Notes:

remove cascade:

  This commit also removes the behaviour which was something like "try
  several statuses before completing". It is the authors view that there
  should be a single workflow per project, and that it can be codified
  through configuration.

  There is no handling for multiple projects.